resolve decimal bug

+++ Onyx/Models/StoredProcedure/EmpLeaveSalaryTicket_Approval_GetRow_Result.cs
@@ -7,8 +7,8 @@
         public string FormatedTransDt { get; set; }
         public string EmpCd { get; set; }
         public string Emp { get; set; }
-        public int LvSalary { get; set; }
-        public int LvTicket { get; set; }
+        public decimal LvSalary { get; set; }
+        public decimal LvTicket { get; set; }
         public int Current_Approval_Level { get; set; }
         public string Current_Approval { get; set; }
         public string Approvals { get; set; }
